requirements.txt
# List of dependencies  

requests==2.31.0    # For Ensembl API calls
pyfaidx==0.8.1.3    # Latest available version of library - for reading genomic coordinates from FASTA
biopython==1.81     # For sequence validation
tensorflow==2.18.0
pandas==2.2.3